chatbot/requirements.txt
# FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# LangChain dependencies
langchain==0.1.1
langchain-community==0.0.10
langchain-huggingface==0.0.6

# ML/AI dependencies
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.0
huggingface-hub==0.19.4
scikit-learn==1.3.2
numpy==1.24.3

# Vector database
faiss-cpu==1.7.4

# Environment and utilities
python-dotenv==1.0.0
pydantic[email]==2.5.0

# Email functionality
email-validator==2.1.0

# File handling
pypdf==6.0.0